+++ providers/implementations/digests/digestcommon.c
