+++ providers/implementations/include/prov/digestcommon.h
